Rename 'hc-upload' or 'Upload' to AvatarUploader and moved 'ImageUploader' to folder 'Uploader'
